Do an audit of hunk mark usage
I realized that after making the hunk 64-bit clean, I had forgotten to
go through and convert all the saved marks to size_t.

+++ libs/gib/gib_init.c
@@ -62,7 +62,7 @@ static void
 GIB_Exec_Override_f (void)
 {
 	char       *f;
-	int         mark;
+	size_t      mark;
 
 	if (Cmd_Argc () != 2) {
 		Sys_Printf ("exec <filename> : execute a script file\n");
